test: 로또 머신 유저 숫자 입력 예외 테스트
